# openapi.yaml
openapi: 3.0.3
info:
  title: HLS Streaming Server API
  description: |
    Complete REST API for managing HLS streaming channels, media libraries, and dynamic playlists.

    ## Features
    - Multi-channel HLS streaming with FFmpeg
    - Dynamic playlist generation via schedule blocks
    - Jellyfin-style library management
    - Media bucket organization
    - EPG/XMLTV generation
    - IPTV M3U playlist export
    - Session-based authentication
    - Virtual time tracking for 24/7 simulation

    ## Authentication
    The API supports two authentication methods:
    1. **API Key** (Legacy): Pass via `X-API-Key` header or `apiKey` query parameter
    2. **Session Token** (Modern): Pass via `Authorization: Bearer <token>` header or `sessionToken` cookie

    Most endpoints require authentication except for streaming and public EPG endpoints.

  version: 1.0.0
  contact:
    name: HLS Streaming Server
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-server.com
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionToken
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
      required:
        - success
...
    Program:
      type: object
      properties:
        start:
          type: string
          format: date-time
        stop:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
paths:
...
  /api/epg/channels/{slug}:
    get:
      tags: [EPG]
      summary: Get EPG for channel
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Channel EPG
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          channel:
                            type: object
                          programs:
                            type: array
                            items:
                              $ref: '#/components/schemas/Program'
  /api/epg/channels/{slug}/current:
    get:
      tags: [EPG]
      summary: Get current and next program
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current program info
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          channel:
                            type: object
                          current:
                            $ref: '#/components/schemas/Program'
                          next:
                            $ref: '#/components/schemas/Program'
  /api/epg/refresh:
    post:
      tags: [EPG]
      summary: Refresh all EPG data
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: EPG refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/epg/channels/{channelId}/regenerate:
    post:
      tags: [EPG]
      summary: Regenerate EPG for channel
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: EPG regenerated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # === LIBRARIES ===
...